requirements.txt
psycopg==3.2.3
pandas==2.2.3
tk==0.1.0
python-dotenv==1.0.1